Fix by comments
